refactor(dialog): component optimization
