+++ packages/components/dialog/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useAction';
+export * from './useSameTarget';
