+++ packages/components/drawer/hooks/index.ts
@@ -1,7 +1,7 @@
 import { computed, ref } from 'vue';
-import { Styles } from '../common';
+import { Styles } from '../../common';
 import { getSizeDraggable, calcMoveSize } from '@tdesign/common-js/drawer/utils';
-import type { TdDrawerProps } from './type';
+import type { TdDrawerProps } from '../type';
 
 export const useDrag = (props: TdDrawerProps) => {
   // 以下为拖拽改变抽屉大小相关 可以抽成hooks
